requirements.txt
python-telegram-bot==20.5
SQLAlchemy==2.0.30
pymysql==1.1.1
apscheduler==3.10.4
python-dotenv==1.0.1
alembic==1.13.1